Change name of folder from util to utils, move it to src folder
